moves wrapper functionality back into exchange.go
